==> 0_prepare_data/02_data_processing/US/01_prepare_tweet_data_US.ipynb <==
# Process US tweet data

Processes and cleans the data, assigns each user/tweets to a county and computes tweet/user indicators

CREATES:
- ID file for all included tweets/users with link to county of tweet and user
- Tweet level data file

################################################################################
# Import modules
################################################################################
import os
import dask
import dask.bag as db
from dask.diagnostics import ProgressBar
import dask.dataframe as dd
import swifter

import pandas as pd
import numpy as np
import statistics as st

import json
import pprint
import geopandas
from shapely.geometry import Polygon
from shapely.geometry import Point
import glob

from matplotlib import pyplot as plt
from collections import Counter

import time
import datetime
import pytz

start = time.time()

# Set utilities
pd.options.display.max_colwidth = 100

################################################################################
# Specify paths
################################################################################
os.chdir("../..")
mode = "full" # sample
if mode == "full":
    input_path = "PATH TO FOLDER"
    output_path = "PATH TO FOLDER"
if mode == "sample":
    input_path = "PATH TO FOLDER"
    output_path = "PATH TO FOLDER"
geo_path = "PATH TO FOLDER"

hours = len(glob.glob(input_path))
print(f"Tweets for {hours} hours = {round(hours/24, 1)} days = {round(hours/24/30,1)} months!")

## 1. Create dask dataframe

tweets = db.read_text(input_path) \
            .filter(lambda x: x != '\n') \
            .filter(lambda x: "created_at" in x) \
            .map(lambda x: x.replace("\n", "")) \
            .map(json.loads)

tweet_info = tweets.map(lambda x: [x["id_str"],
                                   x["created_at"],
                                   x["extended_tweet"]["full_text"] if "extended_tweet" in x else x["text"],
                                   x["user"]["id_str"],
                                   x["user"]["followers_count"],
                                   x["user"]["friends_count"],
                                   x["user"]["statuses_count"],
                                   x["user"]["verified"],
                                   x["user"]["description"],
                                   x["user"]["created_at"],
                                   x["place"]["country_code"] if x["place"]!= None else None,
                                   x["coordinates"]["coordinates"] if x["coordinates"] != None else None,
                                   x["place"]["place_type"] if x["place"] != None else None,
                                   x["place"]["name"] if x["place"] != None else None,
                                   x["place"]["bounding_box"]["coordinates"][0] if x["place"] != None else None,
                                   x["is_quote_status"],
                                   x["user"]["is_translator"],
                                   x["source"],
                                   x["lang"]
                                  ])


tweet_dd = tweet_info.to_dataframe(columns = ["id",
                                              "created_at",
                                              "tweet",
                                              "user_id",
                                              "user_followers",
                                              "user_friends",
                                              "user_statuses",
                                              "user_verified",
                                              "user_description",
                                              "user_created_at",
                                              "country_code",
                                              "coordinates",
                                              "place_type",
                                              "place_name",
                                              "bounding_box",
                                              "is_quote_status",
                                              "is_translator",
                                              "source",
                                              "language"
                                             ])



# Keep only English tweets in the US (and drop translators)
tweet_dd = tweet_dd[(tweet_dd["country_code"] == "US") & (tweet_dd["language"] == "en")]
tweet_dd = tweet_dd[tweet_dd["is_translator"]==False]

tweet_dd = tweet_dd.drop(["is_translator", "country_code", "language"], axis=1)

tweet_dd

## 2. Inspect, clean and prepare georeferenced data

### 2.1 Clean and inspect geodata of tweets

place_types = tweet_dd.place_type.value_counts().compute(scheduler='processes')
print(place_types)

# Check if admin is state or county
tweet_dd[tweet_dd.place_type == "admin"].head() # = state

# Drop country and state level data
tweet_dd = tweet_dd[(tweet_dd["place_type"].isin(["country", "admin"]) == False) | (tweet_dd["coordinates"]).notnull()]
tweet_dd.count().compute(scheduler='processes')

### 2.2 Convert tweet data to geodataframe

# Check and clean place types
df = tweet_dd.compute(scheduler='processes')
df.loc[df["coordinates"].notnull(), "place_type"] = "exact"
df.loc[df["coordinates"].notnull(), "place_name"] = np.nan
print(df.place_type.value_counts())
df.place_type.isna().sum() # Should be 0

# Convert bounding box and exact coordinates to geometry
df["bounding_box"] = df["bounding_box"].apply(Polygon)
df["coordinates"] = df["coordinates"].apply(lambda x: Point(x) if x != None else np.nan)
df["bounding_box"].head()

# Compute best coordinates (exact coords if available, else centroid of bounding box)
df["best_coords"] = df.apply(lambda x: x["coordinates"] if pd.notna(x["coordinates"])
                                     else x["bounding_box"].centroid, axis=1)
df.drop("coordinates", inplace = True, axis = 1)
df["best_coords"].head()

df.loc[df["place_type"] == "exact", "bounding_box"] = np.nan # Set bounding box to na if exact coordinates are available

# Change projection (to make map look better)
df = df.set_geometry('bounding_box')
df.crs = {'init': 'epsg:4326', 'no_defs': True}
df = df.to_crs("EPSG:5071")
for column in ["best_coords", "bounding_box"]:
    print(df[column].head(2))

df = df.set_geometry('best_coords')
df.crs = {'init': 'epsg:4326', 'no_defs': True}
df = df.to_crs("EPSG:5071")
for column in ["best_coords", "bounding_box"]:
    print(df[column].head(2))

# Import shp-file on us counties and change projection
counties = geopandas.read_file(geo_path + r"\cb_2018_us_county_500k.shp") #!!!!!!!!! CHANGED SHP-file
print(counties.crs)
counties = counties.to_crs("EPSG:5071")
print(counties.crs)

# Mark counties where centroid lies in other county
counties["centroid_bb"] = counties["geometry"].envelope.centroid
counties["centroid_in_mun"] = counties["geometry"].contains(counties["centroid_bb"] )
counties["centroid_in_mun"].value_counts()
print(len(counties.loc[counties["centroid_in_mun"] == False, ["NAME", "GEOID"]]))
counties.loc[counties["centroid_in_mun"] == False, ["NAME", "GEOID"]].head()

# Check numer of counties with non-unique names (same names in different states)
print(counties.groupby("NAME")["NAME"].transform("count").value_counts())
print(counties.groupby(["STATEFP", "NAME"])["NAME"].transform("count").value_counts()) # Almost unique within states (6 problem counties)

### 2.3 Join with counties dataset

# Note: First performs different types of joins and then compares and checks consistency to perform final join

# Compute area of bounding box for tweets
print("Tweets:")
df["bounding_box_area"] = df["bounding_box"].area/(1000*1000)
df.loc[df["place_type"] == "exact", "bounding_box_area"] = 0
print(df["bounding_box_area"].describe())

# Compute area of bounding box for counties
print("Counties:")
print((counties["geometry"].envelope.area/(1000*1000)).describe())

df.loc[df["bounding_box_area"] > 2000, "place_name"].value_counts() # Bounding boxes seem to be reasonably precise

# 1. Join if tweet bounding box is within county polygon (works if county level precision)

# Perform within join for polygons
counties.rename({"geometry": "mun_polygon"}, inplace=True, axis=1)
counties = counties.set_geometry('mun_polygon')
df = df.set_geometry('bounding_box')
df = geopandas.sjoin(df, counties[["GEOID", "NAME", "STATEFP", "mun_polygon"]],
                     how='left', op='within')
df.reset_index(drop=True, inplace=True)
df.drop("index_right", inplace=True, axis=1)

# Rename columns
df.rename({"GEOID": "mun_id_within",
           "NAME": "mun_name_within",
           "STATEFP": "state_id_within",
           'centroid_in_mun': "centroid_in_mun_within"}, axis=1, inplace=True)

# Check results
print(df["mun_id_within"].notna().value_counts()) # Only about 1/4 of observations matched
df.loc[df["mun_id_within"].isna(), ["place_type", "place_name"]].value_counts().head(5)

# 2. Join if tweet bounding box contains county polygon (works if only county level precision)

# Perform contains join for polygons
print(len(df))
df = geopandas.sjoin(df, counties[["GEOID", "NAME", "STATEFP", "mun_polygon"]],
                     how='left', op='contains')
df.reset_index(drop=True, inplace=True)
df.drop("index_right", inplace=True, axis=1)
print(len(df)) # Check if it remains the same (if not, some tweet bounding boxes contain several counties)

# Rename columns
df.rename({"GEOID": "mun_id_contains",
           "NAME": "mun_name_contains",
           "STATEFP": "state_id_contains",}, axis=1, inplace=True)

# Check nr of tweets that were merged
print(df["mun_id_contains"].notna().value_counts()) # Very few matches (tweet bounding boxes are not usually county level precision)

# 3. Join with county that contains centroid of bounding box (problem if mun level precision and centroid not in mun)

# Perform within join for centroids
print(len(df))
df = df.set_geometry("best_coords")
df = geopandas.sjoin(df, counties[["GEOID", "NAME", "STATEFP", "centroid_in_mun", "mun_polygon"]],
                     how='left', op='within')
df.reset_index(drop=True, inplace=True)
df.drop("index_right", inplace=True, axis=1)
len(df)

# Rename columns
df.rename({"GEOID": "mun_id_cwithin",
           "NAME": "mun_name_cwithin",
           "STATEFP": "state_id_cwithin",
           'centroid_in_mun': "centroid_in_mun_cwithin"}, axis=1, inplace=True)
df.geometry.name = "best_coords"

# Check merge
print(df["mun_id_cwithin"].notna().value_counts()) # Almost all tweets merged
df.loc[df["mun_id_cwithin"].isna(), ["place_type", "place_name"]].value_counts().head(5) # Mostly coastal places

# 4. Join with county with nearest bounding box centroid (only correct if county level precision)
counties = counties.set_geometry('centroid_bb')
df = df.set_geometry("best_coords")
print(len(df))
df = geopandas.sjoin_nearest(df, counties[["GEOID",
                                       "NAME",
                                       "STATEFP",
                                       "centroid_in_mun",
                                       "centroid_bb"]].set_geometry('centroid_bb'))
df.reset_index(drop=True, inplace=True)
print(len(df))

# Rename columns
df.rename({"GEOID": "mun_id_nearest",
           "NAME": "mun_name_nearest",
           "STATEFP": "state_id_nearest",
           'centroid_in_mun': "centroid_in_mun_nearest"}, axis=1, inplace=True)

# Check merge
print(df["mun_id_nearest"].notna().value_counts()) # Everything merged (by design)

### Check consistency and perform final merge

# Note: For matches that are deemed correct, the respective county IDs are inserted into column GEOID

df.place_type.value_counts()

# Take (1) within, (2) contains matches for polygons and (3) matches if exact/neighborhood/poi coordinates (safe matches)
df["GEOID"] = df.apply(lambda x:
                       x["mun_id_within"] if pd.notna(x["mun_id_within"])
                       else (x["mun_id_contains"] if pd.notna(x["mun_id_contains"])
                       else (x["mun_id_cwithin"] if x["place_type"] in ["exact", "poi", "neighborhood"]
                             else np.nan)),
                       axis=1)
df["GEOID"].notna().value_counts()

# Take cwithin matches (bb centroids within county) if centroid of matched county is within county
print(df["centroid_in_mun_cwithin"].value_counts())
df["GEOID"] = df.apply(lambda x:
                       x["GEOID"] if pd.notna(x["GEOID"])
                       else (x["mun_id_cwithin"] if x["centroid_in_mun_cwithin"] == True
                             else np.nan),
                       axis=1)
df["GEOID"].notna().value_counts() # Very large share of tweets matched

# Take cwithin matches if cwithin and nearest yield same county
print((df["mun_id_cwithin"] == df["mun_id_nearest"]).value_counts())
df["GEOID"] = df.apply(lambda x:
                       x["GEOID"] if pd.notna(x["GEOID"])
                       else (x["mun_id_cwithin"] if x["mun_id_cwithin"] == x["mun_id_nearest"]
                             else np.nan),
                       axis=1)
df["GEOID"].notna().value_counts() # Almost all tweets matched

# Take cwithin  if place_names is equal to county name for (1) cwithin and (2) nearest if equal to nearest match
df["GEOID"] = df.apply(lambda x:
                       x["GEOID"] if pd.notna(x["GEOID"])
                       else (x["mun_id_cwithin"] if x["mun_name_cwithin"] == x["place_name"]
                       else (x["mun_id_nearest"] if x["mun_name_nearest"] == x["place_name"]
                             else np.nan)),
                       axis=1)
df["GEOID"].notna().value_counts()

# Take cwithin if bounding box area is small
df["GEOID"] = df.apply(lambda x:
                       x["GEOID"] if pd.notna(x["GEOID"])
                       else (x["mun_id_cwithin"] if x["bounding_box_area"]<300
                             else np.nan),
                       axis=1)
df["GEOID"].notna().value_counts() # Almost all tweets matched

# Inspect remaining tweets
print(df[df["GEOID"].isna()]["mun_name_cwithin"].notna().value_counts())
df.loc[df["GEOID"].isna() & df["mun_name_cwithin"].notna(), ["place_name", "place_type","mun_name_cwithin", "mun_name_nearest"]].value_counts(dropna=False)

df[df["GEOID"].isna()]["bounding_box_area"].describe()

# Drop if not assigned to any county
print(f"{df['GEOID'].isna().mean()*100} percent of tweets will be dropped") # 0.2 % of tweets dropped
df = df[df['GEOID'].notna()]

# Add name and drop redundant variables
df = df.merge(counties[["NAME", "GEOID"]],  how="left", on = "GEOID")

df.drop(['mun_id_within', 'mun_name_within', 'state_id_within', 'mun_id_contains',
        'mun_name_contains', 'state_id_contains', 'mun_id_cwithin',
        'mun_name_cwithin', 'state_id_cwithin', 'centroid_in_mun_cwithin',
        'index_right', 'mun_id_nearest', 'mun_name_nearest', 'state_id_nearest',
        'centroid_in_mun_nearest','bounding_box_area'], axis=1, inplace=True)

df.columns

### Drop duplicates

# Inspect duplicates
df["dupl_id"] = df.duplicated(subset='id', keep = False)
print(df.dupl_id.sum())
dupls = df[df["dupl_id"]].sort_values(by=["user_id", "id"])
#dupls.head(200).to_excel(output_path + r"\inspect\check_dupls.xlsx") # Seem to be real duplicates, not just duplicated ids

# Drop duplicates
print(len(df))
df.drop_duplicates(subset='id', keep='first', inplace=True)
df.drop(["dupl_id"], axis = 1, inplace =True )
print(len(df))

## 3. Exlude bots

### 3.1 Inspect source of tweets

df.source.value_counts().to_excel(output_path + r"\inspect\sources.xlsx")
df.source.value_counts().head(10)

# Inspect tweets from instagram
df.loc[df.source == """<a href="http://instagram.com" rel="nofollow">Instagram</a>""", "tweet"].head(500).to_excel(
    output_path + r"\inspect\instagram_examples.xlsx")
df[df.source == """<a href="http://instagram.com" rel="nofollow">Instagram</a>"""].head() # Does not look like bots

df[df.source == """<a href="https://www.careerarc.com" rel="nofollow">CareerArc 2.0</a>"""].head()

# Add column to identify tweets that are not posted through a third party API (???)
keep_sources = ["""<a href="http://twitter.com/download/iphone" rel="nofollow">Twitter for iPhone</a>""",
                """<a href="http://twitter.com/download/android" rel="nofollow">Twitter for Android</a>""",
                """<a href="http://instagram.com" rel="nofollow">Instagram</a>""",
                """<a href="http://twitter.com/#!/download/ipad" rel="nofollow">Twitter for iPad</a>"""]

df["nobot"] = False
df.loc[df["source"].isin(keep_sources), "nobot"] = True
df[df["nobot"]].sample(10)

df["nobot"].value_counts()

### 3.2. Inspect users with high number of statuses

df.groupby("nobot")["user_statuses"].describe()

df.loc[(df["user_statuses"] > 200000) & (df["nobot"]),
             ["NAME", "tweet", "user_id", "user_description", "source"]].drop_duplicates(subset="user_id").head(10) # Does not look like bots

### 3.3. Exclude bots using source of tweets

# Keep only: iPhone, Android, Instagram, iPad
df = df[df["nobot"]]
df.drop("nobot", axis=1, inplace=True)

## 4. Compute local time for each tweet

# Inspect format (UCT time)
df["created_at"].head()

# Compute time zone based on coordinates
from tzwhere import tzwhere # Not sure why, but needs to be imported here
tzwhere = tzwhere.tzwhere(forceTZ=True)
df['lon'] = df.to_crs("epsg:4326").best_coords.x
df['lat'] = df.to_crs("epsg:4326").best_coords.y
df["timezone"] = df.apply(lambda x: tzwhere.tzNameAt(x["lat"], x["lon"], forceTZ=True), axis = 1)
print(sum(df["timezone"].isna())) # Should be none
df["timezone"].value_counts().head()

# Compute local time for each tweet
df["created_at"] = pd.to_datetime(df["created_at"], format = '%a %b %d %H:%M:%S +0000 %Y', utc=True)
def tz_func(x):
    return x.dt.tz_convert(x.name).dt.tz_localize(tz=None)
df["local_time"] = df.groupby("timezone")["created_at"].transform(tz_func)
df[["created_at", "timezone", "local_time"]].head()

# Create separate variables for day and time
df["hour"] = df["local_time"].dt.hour
df["day"] = df["local_time"].dt.weekday # 0 = Monday, 6 = Sunday
print(df["hour"].value_counts())
print(df["day"].value_counts())

# Define weekdays
df["weekday"] = df["day"] < 5
df["weekday"].value_counts()

# Distinguish between free time and worktime
df["workhours"] = (df["weekday"]) & (df["hour"] >= 9) & (df["hour"] < 17)
df["workhours"].value_counts()

## 5. Handle users that tweet from different counties

# Inspect nr of counties users tweet from
counties_per_user = df.groupby(["user_id", "GEOID"]).count()
counties_per_user = counties_per_user.groupby("user_id")["lon"].count().sort_values(ascending=False)
print(counties_per_user.describe())
print(len(counties_per_user[counties_per_user>5]))
print(len(counties_per_user[counties_per_user>10]))
print(len(counties_per_user[counties_per_user>20]))
print(len(counties_per_user[counties_per_user>50]))
print(len(counties_per_user[counties_per_user>100]))
print(len(counties_per_user[counties_per_user==1]))
counties_per_user.value_counts(normalize=True).head(10)

# Inspect users who tweet from many counties
df["nr_counties"] = df.groupby("user_id")["GEOID"].transform("nunique") # add variable for nr or counties each user tweets from
if mode == "full":
    df.loc[df["nr_counties"] > 50, ["NAME", "tweet", "user_description"]].sample(20)  # Looks like normal users (e.g. truck drivers)

# Get main county/counties for each user
df["main_county_all"] = df.groupby("user_id")["GEOID"].transform(lambda x: ",".join(st.multimode(x)))
df["nr_main_counties"] = df.groupby("user_id")["GEOID"].transform(lambda x: len(st.multimode(x)))
print(df["main_county_all"].value_counts().head())
print(df["nr_main_counties"].value_counts())
print(df["nr_main_counties"].value_counts(normalize=True))

# Get main county/counties for each user excluding tweets during workhours or weekends (when people might not be at home)
df["county_home"] = df["GEOID"]*1
df["county_home"].loc[(df["weekday"] == False) | (df["workhours"] == True)] = np.nan
df["main_county_home"] = df.groupby("user_id")["county_home"].transform(lambda x: ",".join(
    st.multimode(x[x.notna()])))

df["nr_home_counties"] = df.groupby("user_id")["county_home"].transform(lambda x: len(st.multimode(x[x.notna()])))
print(df["main_county_home"].value_counts().head())
print(df["nr_home_counties"].value_counts())
print(df["nr_home_counties"].value_counts(normalize=True))

df[["main_county_all", "main_county_home", "nr_counties", "nr_home_counties"]].head(20)

# Define var "main_county": main county if unambiguous, else main home county (based on time) if unambiguous, else NaN
df.reset_index(drop=True, inplace=True)
df["main_county"] = df.apply(lambda x:
                                     x["main_county_all"] if x["nr_main_counties"] == 1
                                     else (x["main_county_home"] if x["nr_home_counties"] == 1
                                           else np.nan),
                                     axis=1)
df["main_county"].head()

# Inspect number of ambiguous cases
print(df["main_county"].isna().sum())
print(df["main_county"].isna().mean())

df[["GEOID", "user_id", "main_county_all", "main_county"]].head()

# Get county-user for main county
print("Total nr of counties: ", len(counties["GEOID"].unique()))
print("Nr of counties with at least one tweet: ", len(df["GEOID"].unique()))
print("Nr of counties with at least one tweet as main county: ", len(df["main_county"].unique()))

df.groupby("main_county")["user_id"].count().describe()

## 6. Finalize and export dataset and ID file

### 6.1. Clean and add columns with additional information

df.columns

# Drop redundant columns
df.drop(['main_county_all',
         'nr_main_counties',
         'county_home',
         'main_county_home',
         'nr_home_counties'],
        inplace = True,
        axis = 1)

# Place type and coordinates
df.drop(['best_coords'],
        inplace = True, axis = 1)
df["place_type"].value_counts()

# County name and ID
df.rename({"GEOID":"county_id_tweet",
           "NAME" :"county_name_tweet",
           "main_county": "county_id_user"},
          axis=1, inplace=True)
df = df.merge(counties[["GEOID", "NAME"]],
              left_on="county_id_user",
              right_on="GEOID",
              how="inner")
df.drop("GEOID", inplace=True, axis=1)
df.rename({"NAME": "county_name_user"}, axis=1, inplace=True)

# iPhone users
df["iphone"] = df["source"].isin(["""<a href="http://twitter.com/download/iphone" rel="nofollow">Twitter for iPhone</a>""",
                                  """<a href="http://twitter.com/#!/download/ipad" rel="nofollow">Twitter for iPad</a>"""])
df["iphone"].value_counts()

# Instagram retweets
df["insta"] = df["source"] == """<a href="http://instagram.com" rel="nofollow">Instagram</a>"""
print(df["insta"].value_counts(normalize=True))

# Check Instagram tweets with default retweet text
insta = df[df["source"] == """<a href="http://instagram.com" rel="nofollow">Instagram</a>"""]
print(insta["tweet"].str.startswith("Just posted a photo").value_counts(normalize=True))

# Exclude Instagram tweets with default retweet text
df = df.loc[~((df["source"] == """<a href="http://instagram.com" rel="nofollow">Instagram</a>""") &
              (df["tweet"].str.startswith("Just posted a photo"))), :]

# Length of tweet and user description
df["tweet_length"] = df["tweet"].str.len()
df["user_length"] = df["user_description"].str.len()
df.loc[df["user_length"].isna(), "user_length"] = 0
print(df["tweet_length"].describe())
print(df["user_length"].describe())

# Mobility of users
df.rename({"nr_counties":"user_nr_counties"}, inplace=True, axis=1)
df["user_nr_counties"].describe()

# Order columns
move_cols = ['county_id_tweet', 'county_name_tweet', 'county_id_user', 'county_name_user']
cols  = move_cols + [col for col in df.columns if col not in move_cols]
df = df[cols]
df.columns

# Drop tweets in US territories (Puerto Rico, Virgin Islands)
df["state_id_user"] = df["county_id_user"].str[:2]

df.loc[df["state_id_user"].astype(int) >=60, "state_id_user"].value_counts()

df = df[df["state_id_user"].astype(int)<60] # Drop if not in a US state (but only in territories)

print(df["state_id_user"].value_counts())
print(len(df["state_id_user"].value_counts()))
df.drop(columns="state_id_user", inplace=True)

print(len(df))
df.head()

### 6.2. Export files


# Export ID file
df[["id", "user_id", "county_id_tweet", "county_id_user"]].to_csv(output_path + r"\ids.csv",
                                                                  encoding="utf-8",
                                                                  sep=";",
                                                                  index=False)

# Export full dataset (takes long)
df.to_csv(output_path + r"\tweet_data.csv",
          encoding="utf-8",
          sep=";",
          index=False)

# Print nr of users and tweets in final dataset
print(len(df))
print(df["user_id"].nunique())


df["place_type"].value_counts(normalize=True)
